refactor(gameController): Move update controller to own file
