+++ packages/server/src/modules/games/controllers/index.ts
@@ -1,23 +1,18 @@
 import type { Request, Response } from "express";
 import { z } from "zod";
 import { services } from "../services";
-import { services as playersServices } from "../../players/services";
 import { changeTeamController } from "./changeTeamController";
 import { getPlayedGames } from "../services/getPlayedGames";
 import { getPlayers } from "../services/getPlayers";
 import { registerController } from "./registerController";
-import { dateSchema } from "./types";
 import { removePlayerController } from "./removePlayerController";
 import { removeTeamController } from "./removeTeamController";
 import { putPlayersInTeamsController } from "./putPlayersInTeamsController";
-import { getDefault } from "../services/personalization";
 import { validateProfilesController } from "./validateProfilesController";
 import { updateProfilesController } from "./updateProfilesController";
 import { generateCode } from "../services/utils";
 import { getManyGamesControllers } from "./getManyGamesController";
-import * as playerActionsServices from "../../actions/services/playerActions";
-import * as teamActionsServices from "../../teamActions/services";
-import { batchItems } from "../../../lib/array";
+import { updateGameController } from "./updateGameController";
 
 const crudController = {
   createController,
@@ -31,7 +26,7 @@ const crudController = {
   removeGame,
   removePlayerController,
   removeTeamController,
-  updateGame,
+  updateGameController,
   validateProfilesController,
   updateProfilesController,
 };
@@ -94,58 +89,6 @@ async function getGame(request: Request, response: Response) {
     }
   );
   response.status(200).json({ document });
-}
-
-async function updateGame(request: Request, response: Response) {
-  const paramsSchema = z.object({
-    id: z.string().regex(/^\d+$/).transform(Number),
-  });
-  const bodySchema = z.object({
-    description: z.string().optional(),
-    date: dateSchema.optional(),
-    name: z.string().optional(),
-    status: z.enum(["draft", "ready", "playing", "finished"]).optional(),
-  });
-
-  const { id } = paramsSchema.parse(request.params);
-  const update = bodySchema.parse(request.body);
-
-  if (update?.status === "playing") {
-    await prepareGameForLaunch(id);
-  }
-  const document = await services.update(id, update);
-  response.status(200).json({ document });
-}
-
-async function prepareGameForLaunch(gameId: number) {
-  const BATCH_SIZE = 25;
-
-  const defaultPersonalization = await getDefault();
-  await playersServices.setDefaultProfiles(gameId, defaultPersonalization);
-
-  const game = await services.queries.findUnique({
-    where: {
-      id: gameId,
-    },
-    include: {
-      players: true,
-      teams: true,
-    },
-  });
-
-  await batchItems(game?.players || [], BATCH_SIZE, async (playerBatch) => {
-    const processingPlayerActions = playerBatch.map((p) =>
-      playerActionsServices.getOrCreatePlayerActions(gameId, p.userId)
-    );
-    await Promise.all(processingPlayerActions);
-  });
-
-  await batchItems(game?.teams || [], BATCH_SIZE, async (teamBatch) => {
-    const processingTeamActions = teamBatch.map((t) =>
-      teamActionsServices.getOrCreateTeamActions(t.id)
-    );
-    await Promise.all(processingTeamActions);
-  });
 }
 
 async function removeGame(request: Request, response: Response) {
